Only use one Equalised Fitness metric at a time
Earlier it was possible to use both, which is not the intended purpose of these two metrics. Instead they should only be used one at a time.
When both are assigned, only use EQD, as it has higher threshold for optimal value, so more stricter measure of equalization of fitness for two groups.